+++ .github/workflows/lint.yml
@@ -22,7 +22,7 @@ jobs:
       - name: Install test dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install pre-commit==3.5.0 ruff==0.1.6
+          pip install pre-commit==3.5.0 ruff==0.4.4
 
       - name: Run Ruff
         run: ruff check --output-format=github .
