+++ .pre-commit-config.yaml
@@ -16,7 +16,7 @@ repos:
       - id: requirements-txt-fixer
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.1.6
+    rev: v0.4.4
     hooks:
       - id: ruff
         args: [--fix]
