+++ pyproject.toml
@@ -20,10 +20,9 @@ extend-include = ["*.ipynb"]
 [tool.ruff.lint.per-file-ignores]
 # ignore star and unused imports
 "__init__.py" = ["F403", "F405"]
-# ignore useless expressions used for printing examples
 # ignore module level import not at top of file
-"README.ipynb" = ["B018", "E402"]
-"usage_example.ipynb" = ["B018", "E402"]
+"README.ipynb" = ["E402"]
+"usage_example.ipynb" = ["E402"]
 
 [tool.ruff.lint]
 select = [
